fix #19 : splited core.py in expr.py, group.py, axis.py and array.py
fix #25 : splited test_la.py in common.py, test_axis.py, test_group.py and test_array.py

+ reorganized packages:
- core: expr.py, group.py, axis.py, array.py, session.py, ufuncs.py, ipfp.py
- io: excel.py
- util: misc.py, oset.py
- viewer: api.py, model.py, view.py
- tests: common.py, test_array.py, test_axis.py, test_group.py, test_session.py, test_excel.py, test_ipfp.py
