+++ larray/__init__.py
@@ -1,9 +1,9 @@
 from __future__ import absolute_import, division, print_function
 
 from larray.core import *
-from larray.session import *
-from larray.ufuncs import *
-from larray.excel import open_excel
-from larray.ipfp import ipfp
+from larray.io import *
+from larray.util import *
 from larray.example import load_example_data
 from larray.viewer import *
+
+__version__ = "0.22"
