+++ larray/io/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import absolute_import, division, print_function
+
+from larray.io.excel import open_excel
