+++ setup.cfg
@@ -5,6 +5,6 @@ test=pytest
 testpaths = larray
 # exclude (doc)tests from ufuncs (because docstrings are copied from numpy
 # and many of those doctests are failing
-addopts = -v --doctest-modules --ignore=larray/ufuncs.py
+addopts = -v --doctest-modules --ignore=larray/core/ufuncs.py
 #--maxfail=1 --cov (requires pytest-cov) --pep8 (requires pytest-pep8)
 #pep8maxlinelength = 119
